Moving tests to a more appropriate spot
