+++ app/src/androidTest/java/com/orgzly/android/AllTestSuite.kt
@@ -74,6 +74,8 @@ import org.junit.runners.Suite
         OrgFormatterMiscTest::class,
         OrgFormatterSpeedTest::class,
         OrgFormatterStyleTextTest::class,
+        ContentChangeTitleUpdateViaUpdateNoteTest::class,
+        ContentChangeTitleUpdateViaUpdateNoteContentTest::class,
         StateChangeParentTitleUpdateTest::class,
         UriUtilsTest::class
 )
